Move functions for page 3 to helper module
